+++ spec/lib/terraforming/resource/network_acl_spec.rb
@@ -102,7 +102,8 @@ module Terraforming
         it "should generate tf" do
           expect(described_class.tf(client)).to eq <<-EOS
 resource "aws_network_acl" "hoge" {
-    vpc_id = "vpc-1234abcd"
+    vpc_id     = "vpc-1234abcd"
+    subnet_ids = ["subnet-1234abcd", "subnet-5678efgh"]
 
     ingress {
         from_port  = 0
@@ -119,7 +120,8 @@ resource "aws_network_acl" "hoge" {
 }
 
 resource "aws_network_acl" "fuga" {
-    vpc_id = "vpc-5678efgh"
+    vpc_id     = "vpc-5678efgh"
+    subnet_ids = ["subnet-9012ijkl", "subnet-3456mnop"]
 
     ingress {
         from_port  = 0
@@ -159,6 +161,7 @@ resource "aws_network_acl" "fuga" {
                         "egress.#" => "0",
                         "id" => "acl-1234abcd",
                         "ingress.#" => "1",
+                        "subnet_ids.#" => "2",
                         "tags.#" => "1",
                         "vpc_id" => "vpc-1234abcd",
                       }
@@ -172,6 +175,7 @@ resource "aws_network_acl" "fuga" {
                         "egress.#" => "0",
                         "id" => "acl-5678efgh",
                         "ingress.#" => "1",
+                        "subnet_ids.#" => "2",
                         "tags.#" => "1",
                         "vpc_id" => "vpc-5678efgh",
                       }